requirements.txt
Django==5.0
djangorestframework==3.14.0
drf_spectacular==0.26.5
python-decouple==3.8
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
openai==0.28
django-cors-headers==4.2.0


